requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cssselect==1.3.0
distro==1.9.0
dnspython==2.7.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
googlesearch-python==1.3.0
h11==0.14.0
hdbscan==0.8.40
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.21
langchain-core==0.3.47
langchain-mongodb==0.5.0
langchain-openai==0.3.9
langchain-text-splitters==0.3.7
langsmith==0.3.18
llvmlite==0.44.0
lxml==5.3.1
lxml_html_clean==0.4.1
MarkupSafe==3.0.2
mpmath==1.3.0
narwhals==1.31.0
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.1
numba==0.61.0
numpy==2.1.3
openai==1.68.2
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.4
pyarrow==19.0.1
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pymongo==4.11.3
pynndescent==0.5.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rpds-py==0.23.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
setuptools==77.0.3
sgmllib3k==1.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.39
streamlit==1.43.2
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tinysegmenter==0.3
tldextract==5.1.3
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.50.0
typing_extensions==4.12.2
tzdata==2025.1
umap-learn==0.5.7
urllib3==2.3.0
zstandard==0.23.0

-e .

